+++ src/main/java/bowling/domain/engine/Frame.java
@@ -1,0 +1,8 @@
+package bowling.domain.engine;
+
+public interface Frame {
+
+    void throwBall(PitchResult pitchResult);
+    boolean isEnded();
+
+}
