+++ src/main/java/bowling/domain/engine/Pins.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.engine;
 
 public class Pins {
