+++ src/main/java/bowling/domain/engine/Records.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.engine;
 
 import java.util.ArrayList;
 import java.util.List;
