+++ src/test/java/bowling/domain/engine/PitchResultTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.engine;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
